Hot meals and Queues | codechef
As you know, there is a big problem in the educational system of Chefland - impudent students who cuts in line in the cafeteria. Chef wants to assess the problem. He was able to understand how the queue works:

Each student has value of discontent, which is initially equal to 
0
0;

Initially the queue is empty;

Students arrive at the queue one by one. Each student can either enter the queue in the end, or \textit{cut in line}, by which we mean stand in front of one or more students already in the queue;

When someone cuts in line, the value of discontent of every student who ended up behind the new one is increased by 
1.